requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-dotenv==0.19.0
starlette-context==0.3.3
requests-async==0.6.2
starlette~=0.16.0
http3~=0.6.7
pydantic==1.8.2
pytest==7.1.3